api/requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.6.2
APScheduler==3.11.0
bcrypt==4.2.0
beanie==1.30.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lazy-model==0.2.0
MarkupSafe==3.0.2
motor==3.7.1
msal==1.33.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
requests==2.32.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.39.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.31.1
